# openapi/mocknetpack.yaml
openapi: 3.0.3
info:
  title: MockNetPack API
  version: 0.13.0
  description: |
    MockNetPack 中央 Mock 服务器 API 契约。
    - AI 接入层：为 Agent 工具与脚本/CLI 提供机器原生通道——
      (1) 长效 API Key（POST/GET /api/v1/auth/keys、DELETE /api/v1/auth/keys/{id}）：创建时明文仅返回一次，
      服务端只存 SHA-256 哈希；`Authorization: Bearer <apiKey>` 与登录会话 token 并存，鉴权语义一致；
      (2) 流量数据体验补强：GET /sessions/{id}/traffic 新增 projection=compact（8 字段瘦身视图 +
      每会话自增 seq 增量游标）、since（seq 增量拉取）、method/scheme 过滤，默认（非 compact）响应形状不变，
      人类 Web UI 体验不回归；(3) 错误响应按「给 AI 的指引」：401 提示配置 API Key、404/409/403 说明原因与下一步。
    - 权限与分享（/，v0.11.0）：MockRule 新增 owner（创建者用户名，服务端从当前会话用户写入，
      创建时设）与 updatedBy（最后修改者，每次 PUT 更新）两个纯服务端字段——Web 全量列表下发展示；
      SDK 增量拉取（sinceVersion）只返回 SDK 消费字段（id/method/path/response/enabled/effective），
      绝不携带 owner/updatedBy 等纯服务端/Web-only 字段。权限矩阵：查看规则对设备可见者全部开放
      （设备隔离  不变）；编辑回包/启停 toggle/删除仅规则 owner 与 admin（PUT / DELETE 非 owner
      且非 admin → 403 forbidden，且 403 不泄露规则存在性——对非 admin 未知 ruleId 同样 403）；
      「Mock 此请求」创建新规则所有开发者可用，创建后自己即 owner。存量规则（owner 为空，v0.11.0
      之前创建）仅 admin 可管理，dev 只读 403。分享持久化：ShareSnapshot 随 data.json 落盘，
      重启不失效，7 天 TTL（expiresAt = 创建时刻 + 7 天）不变——过期链接返回 404，janitor 每小时
      清理过期快照（与  48h 会话流量保留并存）。
    - 容量与留存：会话流量上限契约化（--capture-session-max-entries，默认 20000，
      超限丢最旧 rolling window）；服务端过滤下推（GET /sessions/{id}/traffic 新增
      keyword/statusCode/from/to，过滤先于分页）；会话结束后流量保留 48h（--capture-traffic-retention-hours，
      默认 48，自结束时刻起算）供历史日志查询，之后由 janitor 每小时清理；历史日志查询期间保留期内
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /devices/register:
    post:
      tags: [devices]
      summary: 设备注册（幂等 upsert； 起支持配对令牌自动注册）
      description: |
        SDK 启动时调用。did 由 SDK 生成：App 作用域内唯一、App 内持久化（重启不变）、不依赖设备硬件标识。
        相同 (app, did) 重复注册视为同一设备，更新元信息并置为在线（幂等）。
        响应下发服务端心跳配置（间隔/超时为服务端配置项）。
        （v0.8.0）扫码即注册：请求体可携带 pairingToken（Web 二维码中的配对令牌）。令牌有效时，
        未知 (app, did) 直接自动注册（owner=签发令牌的用户，D1），已存在则复用原记录（D7：不新建、
        不改名、owner 为空才补签发用户）；令牌缺失时保持  语义——未知 did 返回 404 device_not_registered。
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '200':
          description: 注册成功（或已存在设备信息更新）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: 配对令牌无效/过期/不匹配该 app（pairing_token_invalid）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 设备未注册（device_not_registered；无配对令牌且 (app, did) 未知）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /devices/{app}/{did}:
    get:
      tags: [devices]
      summary: 设备详情
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/AppPath'
        - $ref: '#/components/parameters/DidPath'
      responses:
        '200':
          description: 设备信息（含当前会话，如有）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [devices]
      summary: 设置设备自定义名称/备注
      operationId: renameDevice
      parameters:
        - $ref: '#/components/parameters/AppPath'
        - $ref: '#/components/parameters/DidPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 64
                  description: 自定义名称（如"张三的 15 Pro"），空串表示清除
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # 7. 配对令牌：扫码连接的基础
  # 设备页为已登录用户签发短时令牌；二维码携带该令牌 + 服务器地址 + appID，
  # SDK 扫码后凭令牌自动注册设备。令牌 10 分钟有效、可复用（同一令牌可注册多台设备）。
  # ==========================================================================
...
  /auth/keys:
    post:
      tags: [auth]
      summary: 创建长效 API Key（明文仅本次返回）
      description: |
        为当前用户签发一个长效 API Key（前缀 mnpk_，服务端存 SHA-256 哈希）。
        明文 key 只在创建响应中出现一次，之后任何接口（含列表）都不再返回；
        store 只落 hash。API Key 无过期（expiresAt 恒为 null，预留扩展）。
        与登录会话 token 并存：`Authorization: Bearer <apiKey>` 与 `<token>` 均可通过鉴权。
      operationId: createApiKey
      responses:
        '201':
          description: 创建成功（key 字段为明文，仅本次返回）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '401':
          description: 缺少或无效的 Bearer 凭证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [auth]
      summary: 当前用户的 API Key 列表（不含明文）
      description: |
        返回当前用户已创建且未吊销的 API Key（id/keyPrefix/createdAt/expiresAt）；
        永不包含 key 明文或哈希。admin 可查看全部用户的 key（owner 校验），dev 仅自己。
      operationId: listApiKeys
      responses:
        '200':
          description: Key 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAPIKeysResponse'
        '401':
          description: 缺少或无效的 Bearer 凭证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/keys/{id}:
    delete:
      tags: [auth]
      summary: 吊销 API Key
      description: |
        吊销指定 API Key（立即失效，不可恢复）。仅 key 的 owner 与 admin 可吊销；
        非 owner 且非 admin 一律 404（不泄露 key 存在性）。成功返回 204。
        幂等：吊销后再次吊销同一 id 返回 404（已不存在）。
      operationId: revokeApiKey
      parameters:
        - $ref: '#/components/parameters/ApiKeyIdPath'
      responses:
        '204':
          description: 已吊销
        '401':
          description: 缺少或无效的 Bearer 凭证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: key 不存在，或当前用户无权吊销（统一 404）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: |
        两种凭证并存：
        - 会话 token：登录（POST /auth/login）签发、服务端持久化；Web 每次请求携带
          `Authorization: Bearer <token>`。有效期 7 天（服务端配置项）；登出即服务端吊销。
        - 长效 API Key：POST /auth/keys 创建，`Authorization: Bearer <mnpk_...>`；无过期，
          明文仅创建时返回一次，服务端只存 SHA-256 哈希（S3：任何日志不落 Authorization/key 明文）。
        两者权限等价（API Key 归属用户角色），401/403 语义一致。
  parameters:
    AppPath:
      name: app
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: App 标识（如 bundle id），路径需 URL 编码
    DidPath:
      name: did
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: 设备标识（SDK 生成，App 作用域内唯一）
...
    ApiKeyIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: API Key ID（创建响应返回的 id，ULID）
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DeviceStatus:
      type: string
      enum: [idle, capturing, offline]
      description: idle=待命（已注册未激活）；capturing=抓包中（有激活会话）；offline=离线（心跳超时）
    Device:
      type: object
      required: [app, did, status, lastSeenAt, registeredAt]
      properties:
        app:
          type: string
          description: App 标识（bundle id）
        appName:
          type: string
          maxLength: 64
          description: |
            App 显示名（SDK 上报 CFBundleDisplayName，如 IntegratingApp；v0.9.0）。
            旧设备/手动注册为空；Web 显示时回退到 app（bundle id）。
        did:
          type: string
          description: 设备标识（SDK 生成）
        name:
          type: string
          nullable: true
          description: 自定义名称（Web 设置）
        owner:
          type: string
          nullable: true
          description: 注册该设备的账号用户名（ Web 手动注册 /  扫码自动注册）；SDK 直接注册的遗留设备为空
        platform:
          type: string
          enum: [ios]
          description: 平台（为 Android 预留， 仅 ios）
        osVersion:
          type: string
          description: 系统版本
        sdkVersion:
          type: string
          description: MockNetPackKit SDK 版本
        appVersion:
          type: string
          description: App 版本
        status:
          $ref: '#/components/schemas/DeviceStatus'
        lastSeenAt:
          type: string
          format: date-time
          description: 最后活跃时间（最近一次心跳）
        registeredAt:
          type: string
          format: date-time
          description: 首次注册时间
        currentSession:
          $ref: '#/components/schemas/CaptureSession'
          description: 当前激活会话（仅 capturing 时有值）
...
    RegisterDeviceRequest:
      type: object
      required: [app, did]
      properties:
        app:
          type: string
          maxLength: 128
        did:
          type: string
          maxLength: 128
          description: SDK 生成；App 作用域内唯一；App 内持久化（重启不变）；不依赖设备硬件标识
        platform:
          type: string
          enum: [ios]
          default: ios
        osVersion:
          type: string
        sdkVersion:
          type: string
        appVersion:
          type: string
        appName:
          type: string
          maxLength: 64
          description: App 显示名（SDK 上报 CFBundleDisplayName，如 IntegratingApp；v0.9.0）。可选；Web 回退到 app。
        pairingToken:
          type: string
          maxLength: 128
          description: |
            配对令牌：Web 二维码中的短时令牌。提供且有效时未知 (app, did)
            自动注册（owner=签发用户，D1），已存在则复用原记录；缺失时保持  语义
            （未知 did 404）。
        deviceName:
          type: string
          maxLength: 64
          description: |
            扫码自动注册时使用的设备显示名（可选）；不提供时服务端回退为
            platform + did 前缀（如 ios·abcdef123456）。已存在设备复用时不生效（不改名，D7）。
...
    RegisterDeviceResponse:
      type: object
      required: [device, serverConfig]
      properties:
        device:
          $ref: '#/components/schemas/Device'
        serverConfig:
          $ref: '#/components/schemas/ServerConfig'
    ServerConfig:
      type: object
      required: [heartbeatIntervalSeconds, heartbeatTimeoutSeconds]
      properties:
        heartbeatIntervalSeconds:
          type: integer
          minimum: 5
          maximum: 300
          default: 20
          description: SDK 心跳间隔（需求默认建议 15~20s，服务端配置项）
        heartbeatTimeoutSeconds:
          type: integer
          minimum: 10
          maximum: 600
          default: 60
          description: 心跳超时阈值：连续超时未收到心跳判定离线并结束会话（服务端配置项）
...
    SessionStatus:
      type: string
      enum: [capturing, ended]
      description: capturing=抓包中；ended=已结束（ 起仅瞬时状态：会话结束即删除记录，不持久化 ended）
    CaptureSession:
      type: object
      required: [id, app, did, status, startedAt]
      properties:
        id:
          type: string
          description: 会话 ID（ULID）
        app:
          type: string
        did:
          type: string
        status:
          $ref: '#/components/schemas/SessionStatus'
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        requestCount:
          type: integer
          default: 0
          description: 会话内请求数（启用后有真实值）
        viewerCount:
          type: integer
          default: 0
          description: 当前打开该设备页面的查看者数（最后一个释放则会话结束）
        retainUntil:
          type: string
          format: date-time
          nullable: true
          description: 结束会话的保留截止时刻
...
    APIKeyView:
      type: object
      required: [id, keyPrefix, username, createdAt]
      description: API Key 列表项——永不包含明文或哈希。
      properties:
        id:
          type: string
          description: Key ID（ULID）
        keyPrefix:
          type: string
          description: 明文 key 前缀（mnpk_ 开头的展示片段，用于区分，不可恢复全文）
        username:
          type: string
          description: 所属用户名
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: 过期时间（当前恒为 null，预留）
    CreateAPIKeyResponse:
      type: object
      required: [id, key, createdAt]
      description: 创建响应——key 字段为明文，仅在创建时返回一次。
      properties:
        id:
          type: string
        key:
          type: string
          description: 明文 API Key（mnpk_ 前缀）；之后任何接口不再返回
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ListAPIKeysResponse:
      type: object
      required: [keys, total]
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/APIKeyView'
        total:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 机器可读错误码（如 device_not_registered）
        message:
          type: string
          description: 人类可读错误信息